Fix HTTPS cookies, fix integer subcategories
First, fix tracking and sending of session cookies.  The key fix in
this change is just send back the simple cookie=val (and not all the
cookie 'attributes' the server sent us).  And to send the cookies back
on a single "Cookie:" header line (separated by ;), instead of adding
a Cookie: header per cookie.

Second, support categories (and subcategories) that contain
integer-only names (like "2010").  The old code assumed that a
digit-only name was a category ID.  This made it impossible to specify
a category whose name only contains digits.  Add explicit
--category-id and --subcategory-id options if someone wants to specify
the ID.  This is a fix for:
	https://github.com/jmrodri/sm-photo-tool/issues#issue/5